requirements.txt

streamlit
langchain
langchain_community
langchain_core
langchain_openai
openai
python-dotenv